+++ Sources/Utils/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.36.0 (July 7th, 2018):
+* [Fix] Fix handling the negative values in all of the type formatters. E.g. [`GUIUtils.AngleType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_AngleType.htm).
+* [Enhancement] Auto update modules that implement [`GUIUtils.IHasContextMenu`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_IHasContextMenu.htm).
+* [Enhancement] Support localization for the `UI_Toggle` attribute: [`GUIUtils.StdSpecTags](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_StdSpecTags.htm).
+* [Enhancement] On language version change, update the events and menus even when the part doesn't have a config: [`GUIUtils.LocalizationLoader](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_LocalizationLoader.htm).
+* [Enhancement] Add `session` group to teh standard persistent groups: [`ConfigUtils.StdPersistentGroups](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ConfigUtils_StdPersistentGroups.htm).
+* [Enhancement] Support `km/s` and `Mm/s` units in `VelocityType`: [`GUIUtils.VelocityType](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_VelocityType.htm).
+
 # 0.35.0 (June 25th, 2018):
 * [Fix] Make a value copy of the persistent fields when copying config from the prefab: [`ConfigAccessor.CopyPartConfigFromPrefab`](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ConfigUtils_ConfigAccessor_CopyPartConfigFromPrefab.htm).
 * [Change] Move specification tag constant from `LocalizationManager` into a specialized class [`GUIUtils.StdSpecTags`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_StdSpecTags.htm).